requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11
groq==0.9.0
python-dotenv==1.0.1
